implement enhancement for calculation

+++ Application/CompanyPosts.Application/Abstraction/ICurrencyRepository.cs
@@ -4,5 +4,5 @@ namespace CompanyPost.Application.Abstraction;
 public interface ICurrencyRepository<T> : IGenericRepository<T>
 where T : BaseEntity, IHasCurrencyAndValue
 {
-    Task<double> SumForCurrency(Currency currency, CancellationToken cancellationToken = default);
+    Task<decimal> SumForCurrency(Currency currency, CancellationToken cancellationToken = default);
 }
